move analyze_sensitivity to base
