+++ paddlex/cv/models/classifier.py
@@ -15,18 +15,15 @@
 from __future__ import absolute_import
 import os.path as osp
 from collections import OrderedDict
-from functools import partial
 import numpy as np
 import paddle
 from paddle import to_tensor
 import paddle.nn.functional as F
-from paddleslim.dygraph import L1NormFilterPruner, FPGMFilterPruner
 from paddlex.utils import logging, TrainingStats
 from paddlex.cv.models.base import BaseModel
 from paddlex.cv.nets.ppcls.modeling import architectures
 from paddlex.cv.nets.ppcls.modeling.loss import CELoss
 from paddlex.cv.transforms import arrange_transforms
-from .slim.prune import _pruner_eval_fn
 
 __all__ = [
     "ResNet18", "ResNet34", "ResNet50", "ResNet101", "ResNet152",
@@ -263,33 +260,6 @@ class BaseClassifier(BaseModel):
         prediction = self._postprocess(prediction, true_topk, self.labels)
 
         return prediction
-
-    def analyze_sensitivity(self,
-                            dataset,
-                            batch_size=8,
-                            criterion='l1_norm',
-                            save_dir='output'):
-        assert criterion in ['l1_norm', 'fpgm'], \
-            "Pruning criterion {} is not supported. Please choose from ['l1_norm', 'fpgm']"
-        arrange_transforms(
-            model_type=self.model_type,
-            transforms=dataset.transforms,
-            mode='eval')
-        self.net.train()
-        inputs = [1] + list(dataset[0][0].shape)
-        if criterion == 'l1_norm':
-            self.pruner = L1NormFilterPruner(self.net, inputs=inputs)
-        else:
-            self.pruner = FPGMFilterPruner(self.net, inputs=inputs)
-
-        sen_file = osp.join(save_dir, 'model.sensi.data')
-        logging.info('Sensitivity analysis of model parameters starts...')
-        self.pruner.sensitive(
-            eval_func=partial(_pruner_eval_fn, self, dataset, batch_size),
-            sen_file=sen_file)
-        logging.info(
-            'Sensitivity analysis is complete. The result is saved at {}.'.
-            format(sen_file))
 
     def _preprocess(self, images, transforms, model_type):
         arrange_transforms(
