+++ paddlex/cv/models/slim/prune.py
@@ -12,7 +12,26 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import paddle
+
 
 def _pruner_eval_fn(model, eval_dataset, batch_size):
     metric = model.evaluate(eval_dataset, batch_size=batch_size)
     return metric[list(metric.keys())[0]]
+
+
+def _pruner_template_input(sample, model_type):
+    if model_type == 'detector':
+        template_input = [{
+            "image": paddle.ones(
+                shape=[1, 3] + list(sample["image"].shape[:2]),
+                dtype='float32'),
+            "im_shape": paddle.full(
+                [1, 2], 640, dtype='float32'),
+            "scale_factor": paddle.ones(
+                shape=[1, 2], dtype='float32')
+        }]
+    else:
+        template_input = [1] + list(sample[0].shape)
+
+    return template_input
